Trigger select events from both controllers
